Separate Gemini OCR and refine functionality
