+++ app/services/ocr/gemini/__init__.py
@@ -1,2 +1,3 @@
-__all__ = ["GeminiOCRService"]
-from .gemini_service import GeminiOCRService
+__all__ = ["GeminiOCRService", "GeminiRefinerService"]
+from .ocr_service import GeminiOCRService
+from .refiner_service import GeminiRefinerService
